moved participant class
