+++ src/ProcessMaker/Nayra/Bpmn/Models/Participant.php
@@ -1,8 +1,10 @@
 <?php
 
-namespace ProcessMaker\Models;
+namespace ProcessMaker\Nayra\Bpmn\Models;
 
+use ProcessMaker\Models\ActivityActivatedEvent;
 use ProcessMaker\Nayra\Bpmn\ParticipantTrait;
+use ProcessMaker\Nayra\Contracts\Bpmn\ActivityInterface;
 use ProcessMaker\Nayra\Contracts\Bpmn\ParticipantInterface;
 
 /**
@@ -12,10 +14,7 @@ use ProcessMaker\Nayra\Contracts\Bpmn\ParticipantInterface;
  */
 class Participant implements ParticipantInterface
 {
-    use ParticipantTrait,
-        LocalFlowNodeTrait,
-        //LocalProcessTrait,
-        LocalPropertiesTrait;
+    use ParticipantTrait;
 
     /**
      * Array map of custom event classes for the bpmn element.
